Add Python script to generate RDK-7 build environment
The `generate-rdk-build-env` script introduced in this patch is intended
to do the job of the `setup.sh` script, but reading all config variables
from a YAML file instead of relying on interaction with the developer.

A Mako template has been used to define the output, and `jsonschema` is
used to validate the config data (yes, this works even though the config
is in YAML:-).  Default output template and config schema files can be
found in the `etc/` directory, but both can be overridden using command
line options if desired/useful.

The generated environment settings are written to `stdout`.  Note that
log/status messages used by the script are written to `stderr`, so the
user should not combine with 2>&1 if e.g. piping to file.